Factor out a util.js from explorer.js.
