+++ frontend/source/js/data-explorer/explorer.js
@@ -1,4 +1,4 @@
-/* global $, window, document, history, event,
+/* global $, window, document, history,
   d3, formdb, XMLSerializer, canvg, wNumb, saveAs, Image, vex */
 
 // wNumb is from nouislider
@@ -10,11 +10,15 @@ import ga from '../common/ga';
 
 import hourglass from '../common/hourglass';
 
-// for IE9: History API polyfill
-const location = window.history.location || window.location;
-// TODO: if location.hash, read that
-// e.g. if an IE9 user sends a link to a Chrome user, they should see the
-// same stuff.
+import {
+  location,
+  getUrlParameterByName,
+  parseSortOrder,
+  arrayToCSV,
+  templatize,
+  getFormat,
+  isNumberOrPeriodKey,
+} from './util';
 
 const MAX_EXPERIENCE = 45;
 const HISTOGRAM_BINS = 12;
@@ -44,33 +48,6 @@ let request;
 let updateCounter = 0;
 let histogramUpdated = false;
 
-
-function getUrlParameterByName(name) {
-  const cleanedName = name.replace(/[[]/, '\\[').replace(/[\]]/, '\\]');
-  const regex = new RegExp(`[\\?&]${cleanedName}=([^&#]*)`);
-  const results = regex.exec(location.search);
-  return results === null ? ''
-    : decodeURIComponent(results[1].replace(/\+/g, ' ')).replace(/[<>]/g, '');
-}
-
-function parseSortOrder(order) {
-  if (!order) {
-    return { key: null, order: null };
-  }
-  const first = order.charAt(0);
-  const sort = { order: '' };
-  switch (first) {
-    case '-':
-      sort.order = first;
-      order = order.substr(1); // eslint-disable-line no-param-reassign
-      break;
-    default:
-      break;
-  }
-  sort.key = order;
-  return sort;
-}
-
 function getExcludedIds() {
   const str = form.get('exclude');
   return str && str.length
@@ -91,18 +68,6 @@ function updateExcluded() {
       : 'none')
     .attr('title', `${rows}: ${excluded.join(', ')}`)
     .text(text);
-}
-
-function arrayToCSV(data) {
-  // turns any array input data into a comma separated string
-  // in use for the education filter
-  Object.keys(data).forEach((k) => {
-    if (Array.isArray(data[k])) {
-      data[k] = data[k].join(','); // eslint-disable-line no-param-reassign
-    }
-  });
-
-  return data;
 }
 
 function excludeRow(id) {
@@ -339,11 +304,6 @@ function updateDescription(res) {
       );
     }
   });
-}
-
-function templatize(str, undef) {
-  const undefFunc = d3.functor(undef);
-  return (d) => str.replace(/{(\w+)}/g, (_, key) => d[key] || undefFunc.call(d, key));
 }
 
 function updatePriceHistogram(data) {
@@ -900,25 +860,6 @@ function setupSortHeaders(headers) {
     .on('click.sort', setSortOrder);
 }
 
-function getFormat(spec) {
-  if (!spec) {
-    return (d) => d;
-  }
-
-  const index = spec.indexOf('%');
-  if (index === -1) {
-    return d3.format(spec);
-  }
-  const prefix = spec.substr(0, index);
-  const format = d3.format(spec.substr(index + 1));
-  return (str) => {
-    if (!str) {
-      return '';
-    }
-    return prefix + format(+str);
-  };
-}
-
 function setupColumnHeader(headers) {
   headers
     .datum(function setupDatum() {
@@ -975,17 +916,6 @@ function histogramToImg(e) {
       },
     });
   }
-}
-
-function isNumberOrPeriodKey(evt) {
-  const charCode = (evt.which) ? evt.which : event.keyCode;
-  if (charCode === 46) {
-    return true;
-  }
-  if (charCode > 31 && (charCode < 48 || charCode > 57)) {
-    return false;
-  }
-  return true;
 }
 
 /*
